module1-task3 - Отредактировал презентер - Удалил логику из events-view
