+++ src/view/events-add-view.js
