+++ src/view/events-edit-view.js
@@ -1,7 +1,7 @@
 import { createElement } from '../render';
 
 export function createEditEventTemplate() {
-  return `<form class="event event--edit" action="#" method="post">
+  return `<li class="trip-events__item"><form class="event event--edit" action="#" method="post">
   <header class="event__header">
     <div class="event__type-wrapper">
       <label class="event__type  event__type-btn" for="event-type-toggle-1">
@@ -153,7 +153,8 @@ export function createEditEventTemplate() {
       <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
     </section>
   </section>
-  </form>`;
+  </form>
+  </li>`;
 }
 
 export default class TripEventsEditView {
