+++ src/view/failed-load-data.js
