+++ src/view/info-view.js
