+++ src/view/loading-view.js
